fix: receive result JSON as object
The `auth/test` endpoint actually returns a JSON object with a single
key "success" that contains a boolean value. Initially the `test`
function was implemented with the assumption that a single boolean
would be the value returned, but this is not the case.

This fix corrects the previous assumption, unwrapping the object and
the value contained at the `"success"` key if it exists.

```
curl --location --request POST 'https://marketplace.tensordock.com/api/v0/auth/test' \
> --header 'Content-Type: application/x-www-form-urlencoded' \
> --data-urlencode 'api_key=API_KEY' \
> --data-urlencode 'api_token=API_TOKEN'
{"success":true}
```

Signed-off-by: Lucas Sta Maria <lucas@priime.dev>